feat: Added NewsArticleAdapter, NewsLoader, QueryUtils and SettingsActivity modules, many resources.
Major update, moved networking from MainActivity to QueryUtils module, added NewsLoader,NewsArcicleAdapter. Module SettingsActivity is added and still need to be customized to reflect choice of news topic. Most of changed modules and resources are in draft state, need to customized them for the news application. Currently the application shows list of current events with section and webTitle, intent for webBrowser works as well with correct URL.